+++ clang/docs/ReleaseNotes.rst
@@ -184,6 +184,7 @@ AST Matchers
 
 clang-format
 ------------
+- Add ``ShortReturnTypeColumn`` option.
 
 libclang
 --------